Reworked draw order. The user have between 0 and 65535 layers to use. Internally components have (1ULL << 48) internal sublayers to use.

+++ src/core/engine_private.c
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #include <string.h>
 #include "utility/log.h"
 #include "entity/entity_private.h"
@@ -18,7 +19,7 @@ void _engine_render_list_clear(EseEngine *engine) {
 }
 
 void _engine_add_texture_to_draw_list(
-    float screen_x, float screen_y, float screen_w, float screen_h, int z_index,
+    float screen_x, float screen_y, float screen_w, float screen_h, uint64_t z_index,
     const char *texture_id, float texture_x1, float texture_y1, float texture_x2, float texture_y2,
     int width, int height,
     void *user_data
@@ -33,7 +34,7 @@ void _engine_add_texture_to_draw_list(
 }
 
 void _engine_add_rect_to_draw_list(
-    float screen_x, float screen_y, int z_index,
+    float screen_x, float screen_y, uint64_t z_index,
     int width, int height, float rotation, bool filled,
     unsigned char r, unsigned char g, unsigned char b, unsigned char a,
     void *user_data
@@ -50,7 +51,7 @@ void _engine_add_rect_to_draw_list(
 }
 
 void _engine_add_polyline_to_draw_list(
-    float screen_x, float screen_y, int z_index,
+    float screen_x, float screen_y, uint64_t z_index,
     const float* points, size_t point_count, float stroke_width,
     unsigned char fill_r, unsigned char fill_g, unsigned char fill_b, unsigned char fill_a,
     unsigned char stroke_r, unsigned char stroke_g, unsigned char stroke_b, unsigned char stroke_a,
