+++ src/entity/entity.h
@@ -2,6 +2,7 @@
 #define ESE_ENTITY_H
 
 #include <stdbool.h>
+#include <stdint.h>
 #include "entity/entity_lua.h"
 #include "scripting/lua_engine_private.h"
 
@@ -27,7 +28,7 @@ typedef struct EseRect EseRect;
  * @param user_data     User-provided callback data
  */
 typedef void (*EntityDrawTextureCallback)(
-    float screen_x, float screen_y, float screen_w, float screen_h, int z_index,
+    float screen_x, float screen_y, float screen_w, float screen_h, uint64_t z_index,
     const char *texture_id, float texture_x1, float texture_y1, float texture_x2, float texture_y2,
     int width, int height,
     void *user_data
@@ -50,7 +51,7 @@ typedef void (*EntityDrawTextureCallback)(
  * @param user_data     User-provided callback data
  */
 typedef void (*EntityDrawRectCallback)(
-    float screen_x, float screen_y, int z_index,
+    float screen_x, float screen_y, uint64_t z_index,
     int width, int height, float rotation, bool filled,
     unsigned char r, unsigned char g, unsigned char b, unsigned char a,
     void *user_data
@@ -76,7 +77,7 @@ typedef void (*EntityDrawRectCallback)(
  * @param user_data     User-provided callback data
  */
 typedef void (*EntityDrawPolyLineCallback)(
-    float screen_x, float screen_y, int z_index,
+    float screen_x, float screen_y, uint64_t z_index,
     const float* points, size_t point_count, float stroke_width,
     unsigned char fill_r, unsigned char fill_g, unsigned char fill_b, unsigned char fill_a,
     unsigned char stroke_r, unsigned char stroke_g, unsigned char stroke_b, unsigned char stroke_a,
